refactor: move the validation errors controller out of the player router
