+++ src/routers/player_router.js
@@ -1,7 +1,7 @@
 import express from "express";
-import { validationResult } from "express-validator";
-import { playerIdSchema, playerSchema } from "../utils/validation_schema";
 import player from "../controllers/player_controller";
+import { playerIdSchema, playerSchema } from "../utils/validation_schema";
+import { playerValidationErrors } from "../controllers/error_controller";
 
 const router = express.Router();
 
@@ -15,13 +15,6 @@ router.put("/", playerSchema, player.update);
 
 router.delete("/:playerId", playerIdSchema, player.delete);
 
-router.use((err, req, res, next) => {
-  const errors = validationResult(req).formatWith(({ param, msg }) => ({
-    [param]: msg
-  }));
-  if (!errors.isEmpty())
-    res.status(405).json(errors.array({ onlyFirstError: true }));
-  else res.status(400).json({ errors: "something went wrong" });
-});
+router.use(playerValidationErrors);
 
 export default router;
